Postfix #3 Memory Manager
In issue#3, the interface for scan.l to malloc new memory is not
suitable, So in the fix, just leave the interface, later i will add
a new memory manage mechanism which like the native memory manage
interface provided like glibc. So that will use TLS to handle the
memory manager of each thread.